Move ComparablePeriod to standalone class
