+++ kie-dmn/kie-dmn-feel/src/main/java/org/kie/dmn/feel/lang/ast/RangeNode.java
@@ -17,23 +17,16 @@
 package org.kie.dmn.feel.lang.ast;
 
 import java.time.Period;
-import java.time.chrono.ChronoPeriod;
-import java.time.chrono.Chronology;
-import java.time.temporal.ChronoUnit;
-import java.time.temporal.Temporal;
-import java.time.temporal.TemporalAmount;
-import java.time.temporal.TemporalUnit;
-import java.util.List;
 
 import org.antlr.v4.runtime.ParserRuleContext;
 import org.kie.dmn.api.feel.runtime.events.FEELEvent.Severity;
 import org.kie.dmn.feel.lang.EvaluationContext;
 import org.kie.dmn.feel.lang.Type;
 import org.kie.dmn.feel.lang.types.BuiltInType;
+import org.kie.dmn.feel.lang.types.impl.ComparablePeriod;
 import org.kie.dmn.feel.runtime.Range;
 import org.kie.dmn.feel.runtime.impl.RangeImpl;
 import org.kie.dmn.feel.util.Msg;
-import org.kie.dmn.feel.util.TypeUtil;
 
 public class RangeNode
         extends BaseNode {
@@ -140,112 +133,6 @@ public class RangeNode
             start = null;
         }
         return start;
-    }
-
-    public static class ComparablePeriod implements Comparable<ChronoPeriod>, ChronoPeriod {
-        private final int left;
-        private final String toStringRep;
-        private final Period period;
-
-        public ComparablePeriod(Period value) {
-            this.period = value;
-            this.left = value.getYears() * 12 + value.getMonths();
-            this.toStringRep = TypeUtil.formatPeriod(value, true);
-        }
-
-        public static ComparablePeriod parse(CharSequence text) {
-            return new ComparablePeriod(Period.parse(text));
-        }
-
-        @Override
-        public int compareTo(ChronoPeriod o) {
-            int right = (int) (o.get(ChronoUnit.YEARS) * 12L + o.get(ChronoUnit.MONTHS));
-            return left - right;
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if ( this == o ) return true;
-            if ( !(o instanceof ComparablePeriod) ) return false;
-
-            ComparablePeriod that = (ComparablePeriod) o;
-
-            return left == that.left;
-        }
-
-        @Override
-        public int hashCode() {
-            return left;
-        }
-
-        @Override
-        public String toString() {
-            return toStringRep;
-        }
-
-        public Period asPeriod() {
-            return period;
-        }
-
-        public static ComparablePeriod ofMonths(int months) {
-            Period p = Period.ofMonths(months);
-            return new ComparablePeriod(p);
-        }
-
-        public static long toTotalMonths(ChronoPeriod left) {
-            return left.get(ChronoUnit.YEARS) * 12L + left.get(ChronoUnit.MONTHS);
-        }
-
-        @Override
-        public long get(TemporalUnit unit) {
-            return period.get(unit);
-        }
-
-        @Override
-        public List<TemporalUnit> getUnits() {
-            return period.getUnits();
-        }
-
-        @Override
-        public Chronology getChronology() {
-            return period.getChronology();
-        }
-
-        @Override
-        public ChronoPeriod plus(TemporalAmount amountToAdd) {
-            return new ComparablePeriod(period.plus(amountToAdd));
-        }
-
-        @Override
-        public ChronoPeriod minus(TemporalAmount amountToSubtract) {
-            return new ComparablePeriod(period.minus(amountToSubtract));
-        }
-
-        @Override
-        public ChronoPeriod multipliedBy(int scalar) {
-            return new ComparablePeriod(period.multipliedBy(scalar));
-        }
-
-        @Override
-        public ChronoPeriod normalized() {
-            return new ComparablePeriod(period.normalized());
-        }
-
-        @Override
-        public Temporal addTo(Temporal temporal) {
-            return period.addTo(temporal);
-        }
-
-        @Override
-        public Temporal subtractFrom(Temporal temporal) {
-            return period.subtractFrom(temporal);
-        }
-
-        public static ComparablePeriod of(int years, int months, int days) {
-            Period p = Period.of(years, months, days);
-            return new ComparablePeriod(p);
-        }
-
     }
 
     @Override
